Remove rotating stuff
