+++ src/main/java/software/bernie/techarium/block/arboretum/ArboretumMaster.java
@@ -14,12 +14,12 @@ import net.minecraft.util.math.shapes.ISelectionContext;
 import net.minecraft.util.math.shapes.VoxelShape;
 import net.minecraft.world.IBlockReader;
 import net.minecraft.world.World;
-import software.bernie.techarium.block.base.MachineBlockRotation;
+import software.bernie.techarium.block.base.MachineBlock;
 import software.bernie.techarium.tile.arboretum.ArboretumTile;
 import software.bernie.techarium.trait.block.BlockBehaviours;
 import software.bernie.techarium.util.BlockRegion;
 
-public class ArboretumMaster extends MachineBlockRotation<ArboretumTile> {
+public class ArboretumMaster extends MachineBlock<ArboretumTile> {
 
     public ArboretumMaster() {
         super(BlockBehaviours.arboretum, AbstractBlock.Properties.of(Material.METAL));
