+++ src/main/java/software/bernie/techarium/block/exchangestation/ExchangeStationBlock.java
@@ -14,11 +14,11 @@ import net.minecraft.util.math.shapes.ISelectionContext;
 import net.minecraft.util.math.shapes.VoxelShape;
 import net.minecraft.world.IBlockReader;
 import net.minecraft.world.World;
-import software.bernie.techarium.block.base.MachineBlockRotation;
+import software.bernie.techarium.block.base.MachineBlock;
 import software.bernie.techarium.tile.exchangestation.ExchangeStationTile;
 import software.bernie.techarium.trait.block.BlockBehaviours;
 
-public class ExchangeStationBlock extends MachineBlockRotation<ExchangeStationTile> {
+public class ExchangeStationBlock extends MachineBlock<ExchangeStationTile> {
 
     public ExchangeStationBlock() {
         super(BlockBehaviours.exchangeStation, AbstractBlock.Properties.of(Material.METAL));
