+++ src/main/java/software/bernie/techarium/block/gravmagnet/GravMagnetBlock.java
@@ -12,11 +12,11 @@ import net.minecraft.state.StateContainer;
 import net.minecraft.state.properties.BlockStateProperties;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
-import software.bernie.techarium.block.base.MachineBlockRotation;
+import software.bernie.techarium.block.base.MachineBlock;
 import software.bernie.techarium.tile.gravmagnet.GravMagnetTile;
 import software.bernie.techarium.trait.block.BlockBehaviours;
 
-public class GravMagnetBlock extends MachineBlockRotation<GravMagnetTile> {
+public class GravMagnetBlock extends MachineBlock<GravMagnetTile> {
 
 	public static final BooleanProperty POWERED = BlockStateProperties.POWERED;
