+++ src/main/java/software/bernie/techarium/tile/base/MachineTileBase.java
@@ -10,7 +10,7 @@ import net.minecraft.tileentity.TileEntity;
 import net.minecraft.tileentity.TileEntityType;
 import net.minecraft.util.ActionResultType;
 import net.minecraft.util.Direction;
-import software.bernie.techarium.block.base.MachineBlockRotation;
+import software.bernie.techarium.block.base.MachineBlock;
 import software.bernie.techarium.machine.sideness.FaceConfig;
 import software.bernie.techarium.machine.sideness.Side;
 
@@ -33,8 +33,8 @@ public abstract class MachineTileBase extends TileEntity {
     public Direction getFacingDirection() {
         assert this.level != null;
         BlockState state = this.level.getBlockState(this.worldPosition);
-        if (state.getBlock() instanceof MachineBlockRotation) {
-        	DirectionProperty direction = ((MachineBlockRotation) state.getBlock()).getDirectionProperty();
+        if (state.getBlock() instanceof MachineBlock) {
+        	DirectionProperty direction = ((MachineBlock) state.getBlock()).getDirectionProperty();
         	return state.getValue(direction);
         }
         return Direction.NORTH;
